Create Cosmalogan or YiazmatMaker; Cosmalog.json
Uh, so i'm a bit new to coding stuff in JSON, aka very new. I'd like to see if you would accept this race as a sort of offering. I apologize if you find this absolutely abhorrent. I sorta copied the frame code from a separate race (Awakened Skeleton), and tried to put in MY informations.